Move AIInterface.{h,cpp} into AI client directory
