--- AI/CMakeLists.txt
@@ -1,6 +1,0 @@
-target_sources(freeorionca
-    PUBLIC
-        ${CMAKE_CURRENT_LIST_DIR}/AIInterface.h
-    PRIVATE
-        ${CMAKE_CURRENT_LIST_DIR}/AIInterface.cpp
-)
