+++ client/AI/AIInterface.cpp
@@ -1,24 +1,24 @@
 #include "AIInterface.h"
 
-#include "../network/ClientNetworking.h"
-#include "../client/AI/AIClientApp.h"
+#include "../../network/ClientNetworking.h"
+#include "AIClientApp.h"
 
-#include "../universe/Universe.h"
-#include "../universe/UniverseObject.h"
-#include "../universe/Planet.h"
-#include "../universe/Fleet.h"
-#include "../universe/Ship.h"
-#include "../universe/ShipDesign.h"
-#include "../universe/System.h"
-#include "../universe/Tech.h"
-#include "../universe/Enums.h"
-#include "../Empire/Empire.h"
+#include "../../universe/Universe.h"
+#include "../../universe/UniverseObject.h"
+#include "../../universe/Planet.h"
+#include "../../universe/Fleet.h"
+#include "../../universe/Ship.h"
+#include "../../universe/ShipDesign.h"
+#include "../../universe/System.h"
+#include "../../universe/Tech.h"
+#include "../../universe/Enums.h"
+#include "../../Empire/Empire.h"
 
-#include "../util/Directories.h"
-#include "../util/Logger.h"
-#include "../util/OptionsDB.h"
-#include "../util/Order.h"
-#include "../util/OrderSet.h"
+#include "../../util/Directories.h"
+#include "../../util/Logger.h"
+#include "../../util/OptionsDB.h"
+#include "../../util/Order.h"
+#include "../../util/OrderSet.h"
 
 #include <boost/python/str.hpp>
 #include <boost/uuid/random_generator.hpp>
