+++ client/AI/CMakeLists.txt
@@ -1,7 +1,9 @@
 target_sources(freeorionca
     PUBLIC
         ${CMAKE_CURRENT_LIST_DIR}/AIClientApp.h
+        ${CMAKE_CURRENT_LIST_DIR}/AIInterface.h
     PRIVATE
         ${CMAKE_CURRENT_LIST_DIR}/AIClientApp.cpp
+        ${CMAKE_CURRENT_LIST_DIR}/AIInterface.cpp
         ${CMAKE_CURRENT_LIST_DIR}/camain.cpp
 )
